Add a unit test for event_base_dump_events()
This function uses a C program to generate its output, and then uses a
Python program to check it for correctness.  On systems without
Python, we just make sure that the C program doesn't crash.

It's likely that we should be requiring some particular python version.
This is an alpha, though: I'm sure somebody will tell us which.