Fix event_dlist definition when sys/queue not included
This header adds TAILQ_HEAD and TAILQ_ENTRY if sys/queue is not included.
There is a similar macro that adds LIST_ENTRY but not LIST_HEAD, even though
LIST_HEAD is used later.

This change pulls in the correct definition (swiped from sys/queue) for
LIST_HEAD and cleans up the one spot where it is used.  The change can be
tested by adding

	#undef LIST_HEAD
	#undef LIST_ENTRY

right before the #ifndef checks and removing

	#define EVENT_DEFINED_LISTENTRY_
	#define EVENT_DEFINED_LISTHEAD_

so that the macros persist later.