evdns: breaking the event loop if there is no requests left.
Here is the brief description of problem:
When you are use evdns to resolve domains to IP adresses (see
./sample/dns-example) you loop never returns from event_base_dispatch(),
and because of this the program will never terminated.

It is not a regress, I'v tested ae5fbf492f (the initial commit of
dns-example sample) and it is also has the same problem.

With "active_requests" field in "nameserver" struct, we can
track active connections, and del/add event when we need it.