Add support for systemd watchdog protocol
This permit to directly have all servers using libevent
to be aware of systemd watchdog monitoring protocol.
It work by sending a message over a specific FD, with
sd_notify.

Unless explicitely enabled by the configuration of the systemd unit,
nothing will be done, so it is safe to enable by default. There is
also protection against accidental triggering with check of the PID
supervised by systemd.

This feature requires at least the version 209 of systemd
due to sd_watchdog_enabled function.