rework libevent linkage
This changes libevent.so so that it doesn't contain any code.  Instead,
it links against the existing core/extra libraries.  At runtime, the
ldso will load those, so the ABI should stay the same.

However, at link time, trying to use -levent with newer linkers might
have trouble.  They will event require you to explicitly link against
the libraries you use and not allow indirect references.  This only
impacts people who hardcode -levent directly rather than using the
pkg-config file, so this shouldn't be a big deal.  If anything, it
encourages people to switch to the pkg-config file.

There's a similar issue for libevent.a as it is empty.  If you use
libtool or pkg-config, things will still work, but if you try to use
-static -levent directly, things will fail.