Replace minheap with William Ahern's timeout code.
There are places where we could use this code more efficiently, and
places where I hacked up the code in a bit of an ugly fashion, but
this should be suitable for a proof-of-concept.

This patch also removes the common_timeout code, on the theory that
it's pointless with this new backend which ought to perform so much
better.

The "#ifdef XXXX" chunks are stuff I can't emulate with current
timeout.c, and I'll have to patch it in.