Fix race condition when deleting a running event
When deleting a running event from a foreign event base, we wait on
a condition variable. In a rare case, the foreign event base may
notify us, then enters another events dispatching, and fires the
event again. However, when we are awake, we don't check if the
event is fired again, and leave the event running, and return.
We may free the resources, which maybe accessed by the event.
Finally, the event may touch the freed memory, IOW, use-after-free
happens.