Remove OpenSSL paragram from README
Because it is mauvais ton to use binaries instead of normal packages
(like apt-get in debian, pacman in arch, and others).

Plus that link was borken and according to [1] OpenSSL do not ship
binaries officially.

And personally I don't think that this is not obvious that you need
openssl libraries to build libevent with it's support, and BTW you need
headers too (of course).

[1]: https://www.openssl.org/community/binaries.html
Fixes: #562

(cherry picked from commit 62866f98a5979e3b9ee8db0287be79749fcf7d6a)