Immediately stop trying to accept more connections if listener disabled
This is a refined version of the logic previously in #578

The rationale is that the consumer of sockets may wish to temporarily
delay accepting for some reason (e.g. being out of file-descriptors).
The kernel will then queue them up. The kernel queue is bounded and
programs like NodeJS will actually try to quickly accept and then close
(as the current behaviour before this PR).

However, it seems that libevent should allow the user to choose whether
to accept and respond correctly if the listener is disabled.