Mitigating numeric type issues found by Veracode
A Veracode (https://www.veracode.com/) scan yielded the following two issues that seem to be valid (although relatively minor) in the code:

Numeric Truncation Error - event.c:2320

Description: There may be a variable truncation in this expression. The maximum or minimum value of the right hand side may be larger or smaller than the maximum or minimum value the variable type of the left hand side can hold. Numeric truncation can also occur when the compiler adheres to integer promotion rules in an arithmetic operation and downcasts the result. Integer truncations can cause security vulnerabilities such as buffer overflows and data corruption, depending on how the numeric value is used.

Remediation: Ensure that no casts, implicit or explicit, take place that move from a larger size primitive to a smaller size primitive.

Signed to Unsigned Conversion Error - buffer.c:1623

Description: This assignment creates a type mismatch by populating an unsigned variable with a signed value. The signed integer will be implicitly cast to an unsigned integer, converting negative values into positive ones. If an attacker can control the signed value, it may be possible to trigger a buffer overflow if the value specifies the length of a memory write.

Remediation: Do not rely on implicit casts between signed and unsigned values because the result can take on an unexpected value and violate weak assumptions made elsewhere in the program.