Fix out-of-tree builds
evconfig-private.h was being included by make dist. Since, for the quote
form of the include directive, headers in the directory of the file with
the #include line have priority over those named in -I options, the copy
of evconfig-private.h from the source directory had priority over the
one in the build directory.