event_exit and fatal_fn from tests: use _exit instead of exit
With multiple threads, if event_exit is called (via
EVUTIL_ASSERT -> event_errx for example), it calls exit() (or fatal_fn which
in the test suite calls exit) which unloads shared libs before exiting. If
another thread is in one of these libs, it can crash and trigger false issues.

Saw this with

  CC="-fsanitize=address" make verify -j8

in dns/getaddrinfo_race_gotresolve_test, which reaches a check fail in

  0  EVUTIL_ASSERT(lock->count == 0);
  1  debug_lock_free()
  2  EVTHREAD_FREE_LOCK(base->th_base_lock, 0);
  3  event_base_free_(struct event_base *base, int run_finalizers)

because lock->count is 1, held_by the other thread which is in
event_base_dispatch().

The visible result is that ASAN reports a heap-use-after-free on the
other thread.